Milestone_4/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.28.0
pydantic>=2.10.0
pydantic-settings>=2.5.0
langgraph>=0.2.0
langchain>=0.2.0
python-dotenv>=1.0.0
duckduckgo-search>=5.0.0
google-generativeai>=0.3.0
